Made it so internal API requests wouldn't make it into access log.
There is a new ENV variable WS_API_LOG_INTERNAL to enable it if you desire, it's mostly helpful for development.
Internal API requests no longer require API key. Migrated system:env to use the API in effort to reduce code duplicating
There going to be more code cleaning later on once we migrate the majority of the code to the API.